requirements.txt
ultralytics==8.2.0
torch==2.2.0
torchvision==0.17.0
streamlit==1.32.0
pillow
numpy
opencv-python-headless
